requirements.txt
requests
flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv